fix: cyclic dependencies of @modern-js/core and @moden-js/webpack
