+++ .changeset/honest-beds-hang.md
@@ -1,0 +1,10 @@
+---
+'@modern-js/core': patch
+'@modern-js/plugin-bff': patch
+'@modern-js/plugin-garfish': patch
+'@modern-js/plugin-storybook': patch
+'@modern-js/webpack': patch
+'@modern-js/utils': patch
+---
+
+fix: cyclic dependencies of @modern-js/core and @moden-js/webpack
