+++ packages/cli/plugin-bff/package.json
@@ -74,7 +74,6 @@
     "@modern-js/plugin-analyze": "workspace:*",
     "@modern-js/runtime": "workspace:*",
     "@modern-js/server-core": "workspace:*",
-    "@modern-js/webpack": "workspace:*",
     "@modern-js/types": "workspace:*",
     "@scripts/build": "workspace:*",
     "@scripts/jest-config": "workspace:*",
