--- packages/cli/webpack/src/compiled.ts
@@ -1,2 +1,0 @@
-export { default as WebpackChain } from '../compiled/webpack-chain';
-export { merge as webpackMerge } from '../compiled/webpack-merge';
